feed-provider/requirements.txt
paho-mqtt
redis
marshmallow-dataclass
fastapi
uvicorn[standard]